pgrep, pkill: Unify arg lists
Most of the process matching arguments are same between pkill and pgrep.
Currently pgrep has more complete list of possible arguments (with
proper types), so copy over the improvements to pgrep.

However, leave arg!()s that are actually not implemented commented out,
so it's easier to see what the uutils version actually supports.