process: Fix tty resolving
Current tty resolution can give wrong result e.g. if the process has
closed all fds of its controlling terminal. The proper way to do it
seems to be based on field in /proc/<pid>/stat, which gives a device
number of the tty. Converting that to a device path requires parsing
/proc/tty/drivers (according to strace of upstream ps).

After this `ps`, `ps -x` and `ps -a` give same processes as upstream ps.